'fix' WFS maxfeature and filter issue
