+++ modules/unsupported/wfs-ng/src/main/java/org/geotools/data/wfs/internal/WFSClient.java
@@ -139,11 +139,6 @@ public class WFSClient extends AbstractOpenWebService<WFSGetCapabilities, QName>
 
         WFSStrategy strategy = null;
 
-        // only one 2.0.0 strategy!
-        if (Versions.v2_0_0.equals(capsVersion)) {
-            strategy = new StrictWFS_2_0_Strategy();
-        }
-
         // override
         if (!"auto".equals(override)) {
             if (override.equalsIgnoreCase("geoserver")) {
@@ -169,6 +164,7 @@ public class WFSClient extends AbstractOpenWebService<WFSGetCapabilities, QName>
         }
 
         // auto detection
+
         if (strategy == null) {
             // look in comments for indication of CubeWerx server
             NodeList childNodes = capabilitiesDoc.getChildNodes();
@@ -219,10 +215,10 @@ public class WFSClient extends AbstractOpenWebService<WFSGetCapabilities, QName>
                     && uri.contains("geoserver")
                     && !Versions.v2_0_0.equals(capsVersion)) {
                 strategy = new GeoServerPre200Strategy();
-            } else if (uri.toLowerCase().contains("/ArcGIS/services/")
+            } else if (uri.toLowerCase().contains("/arcgis/services/")
                     && !Versions.v2_0_0.equals(capsVersion)) {
                 strategy = new ArcGisStrategy_1_X();
-            } else if (uri.toLowerCase().contains("/ArcGIS/services/")
+            } else if (uri.toLowerCase().contains("/arcgis/services/")
                     && Versions.v2_0_0.equals(capsVersion)) {
                 strategy = new ArcGisStrategy_2_0();
             } else if (uri.contains("mapserver") || uri.contains("map=")) {
@@ -236,6 +232,8 @@ public class WFSClient extends AbstractOpenWebService<WFSGetCapabilities, QName>
                 strategy = new StrictWFS_1_x_Strategy(Versions.v1_0_0);
             } else if (Versions.v1_1_0.equals(capsVersion)) {
                 strategy = new StrictWFS_1_x_Strategy(Versions.v1_1_0);
+            } else if (Versions.v2_0_0.equals(capsVersion)) {
+                strategy = new StrictWFS_2_0_Strategy();
             } else {
                 throw new IllegalArgumentException("Unsupported version: " + capsVersion);
             }
@@ -256,7 +254,7 @@ public class WFSClient extends AbstractOpenWebService<WFSGetCapabilities, QName>
     }
 
     public boolean canLimit() {
-        return true;
+        return getStrategy().canLimit();
     }
 
     public boolean canFilter() {
@@ -322,10 +320,10 @@ public class WFSClient extends AbstractOpenWebService<WFSGetCapabilities, QName>
     protected Response internalIssueRequest(Request request) throws IOException {
         Response response;
         try {
-            LOGGER.fine(request.getFinalURL().toString());
+            LOGGER.fine("FinalURL: " + request.getFinalURL().toString());
             if (LOGGER.isLoggable(Level.FINE) && request.requiresPost()) {
                 String msg = request.getPostContentType();
-                LOGGER.fine(msg);
+                LOGGER.fine("Post message: " + msg);
             }
             response = super.internalIssueRequest(request);
         } catch (ServiceException e) {
